Add new flags for part extract
